requirements.txt
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
tqdm>=4.60.0
numba>=0.54.0
jax>=0.2.24
jaxlib>=0.1.75
deap>=1.3.1
pygad>=2.16.3
plotly>=5.3.1
streamlit>=1.0.0
scipy>=1.7.0
Cython>=0.29.24
noise>=1.2.2  # For terrain generation
pyobjc-framework-Metal>=8.5  # For Metal acceleration on macOS
pyobjc-framework-Cocoa>=8.5
pyobjc-framework-Quartz>=8.5